refactor(selector): rename selector to resource-selector
