+++ pkg/controller/blockdeviceclaim/blockdeviceclaim_controller.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	controller2 "github.com/openebs/node-disk-manager/cmd/ndm_daemonset/controller"
-	"github.com/openebs/node-disk-manager/pkg/selector/blockdeviceselect"
-	"github.com/openebs/node-disk-manager/pkg/selector/verify"
+	"github.com/openebs/node-disk-manager/pkg/resourceselector/blockdeviceselect"
+	"github.com/openebs/node-disk-manager/pkg/resourceselector/verify"
 	"github.com/openebs/node-disk-manager/pkg/util"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/reference"
