+++ pkg/resourceselector/verify/resourceverify.go
